vecid use type long

+++ src/main/java/com/guokr/simbase/SimCommand.java
@@ -18,7 +18,7 @@ public abstract class SimCommand {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
-    public void invoke(SimEngine engine, String arg1, int arg2, SimCallback callback) {
+    public void invoke(SimEngine engine, String arg1, long arg2, SimCallback callback) {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
@@ -42,7 +42,7 @@ public abstract class SimCommand {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
-    public void invoke(SimEngine engine, String arg1, String arg2, int arg3, SimCallback callback) {
+    public void invoke(SimEngine engine, String arg1, String arg2, long arg3, SimCallback callback) {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
@@ -62,27 +62,27 @@ public abstract class SimCommand {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
-    public void invoke(SimEngine engine, String arg1, int arg2, String arg3, SimCallback callback) {
+    public void invoke(SimEngine engine, String arg1, long arg2, String arg3, SimCallback callback) {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
-    public void invoke(SimEngine engine, String arg1, int arg2, int arg3, SimCallback callback) {
+    public void invoke(SimEngine engine, String arg1, long arg2, long arg3, SimCallback callback) {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
-    public void invoke(SimEngine engine, String arg1, int arg2, float arg3, SimCallback callback) {
+    public void invoke(SimEngine engine, String arg1, long arg2, float arg3, SimCallback callback) {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
-    public void invoke(SimEngine engine, String arg1, int arg2, String[] arg3, SimCallback callback) {
+    public void invoke(SimEngine engine, String arg1, long arg2, String[] arg3, SimCallback callback) {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
-    public void invoke(SimEngine engine, String arg1, int arg2, int[] arg3, SimCallback callback) {
+    public void invoke(SimEngine engine, String arg1, long arg2, int[] arg3, SimCallback callback) {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
-    public void invoke(SimEngine engine, String arg1, int arg2, float[] arg3, SimCallback callback) {
+    public void invoke(SimEngine engine, String arg1, long arg2, float[] arg3, SimCallback callback) {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
@@ -90,7 +90,7 @@ public abstract class SimCommand {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
-    public void invoke(SimEngine engine, String arg1, float arg2, int arg3, SimCallback callback) {
+    public void invoke(SimEngine engine, String arg1, float arg2, long arg3, SimCallback callback) {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
@@ -110,11 +110,11 @@ public abstract class SimCommand {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
-    public void invoke(SimEngine engine, String arg1, int arg2, String arg3, int arg4, SimCallback callback) {
+    public void invoke(SimEngine engine, String arg1, long arg2, String arg3, long arg4, SimCallback callback) {
         throw new SimCommandException("Wrong parameters passed in");
     }
 
-    public void invoke(SimEngine engine, String arg1, int arg2, String arg3, int[] arg4, SimCallback callback) {
+    public void invoke(SimEngine engine, String arg1, long arg2, String arg3, long[] arg4, SimCallback callback) {
         throw new SimCommandException("Wrong parameters passed in");
     }
